+++ superwall/src/main/java/com/superwall/sdk/dependencies/DependencyContainer.kt
@@ -33,6 +33,7 @@ import com.superwall.sdk.models.events.EventData
 import com.superwall.sdk.models.paywall.Paywall
 import com.superwall.sdk.models.product.ProductVariable
 import com.superwall.sdk.network.Api
+import com.superwall.sdk.network.JsonFactory
 import com.superwall.sdk.network.Network
 import com.superwall.sdk.network.device.DeviceHelper
 import com.superwall.sdk.network.device.DeviceInfo
@@ -101,7 +102,8 @@ class DependencyContainer(
     ConfigManager.Factory,
     AppSessionManager.Factory,
     DebugView.Factory,
-    JavascriptEvaluator.Factory {
+    JavascriptEvaluator.Factory,
+    JsonFactory {
     var network: Network
     override lateinit var api: Api
     override lateinit var deviceHelper: DeviceHelper
@@ -520,8 +522,14 @@ class DependencyContainer(
                             ioScope
                                 .async {
                                     val sandbox =
-                                        JavaScriptSandbox.createConnectedInstanceAsync(context).await()
-                                    SandboxJavascriptEvaluator(sandbox, this@DependencyContainer, storage)
+                                        JavaScriptSandbox
+                                            .createConnectedInstanceAsync(context)
+                                            .await()
+                                    SandboxJavascriptEvaluator(
+                                        sandbox,
+                                        this@DependencyContainer,
+                                        storage,
+                                    )
                                 }.await()
 
                         WebView.getCurrentWebViewPackage() != null ->